+++ CHANGELOG.md
@@ -12,4 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Firebase account creation and authentication
 - Retrieval of locations, climbs and attempts from firebase
 - Views for locations and climbs
-- Location and grade set creation
+- Location, sections and grade set creation